fixed: 修复警告：warning C4244: 'initializing': conversion from 'double' to 'float', possible loss of data

+++ Hazel/src/Hazel/Application/Application.cpp
@@ -45,7 +45,7 @@ namespace Hazel
     {
         while (m_Running)
         {
-            float    time     = glfwGetTime(); // GetTime();
+            float    time     = (float)glfwGetTime(); // GetTime();
             Timestep timestep = time - m_LastFrameTime;
             m_LastFrameTime   = time;
